feat(swift): Parse semgrep_ellipsis
This introduces a `semgrep_ellipsis` rule that is a valid expression. It
conflicts with the existing unbounded range operator. This conflict is
resolved as described in the comment.

Test plan:

Existing tests ensure that unbounded ranges are parsed as they used to
be. The added tests make that clear, and also demonstrate a case where
an ellipsis is now parsed. That example used to be a parse error.