feat(ts): Add ellipsis and top level switch case parsing
We recently received a feature request for allowing standalone switch
cases as patterns, e.g. `case 5: ...`. This seems useful. There is
currently no precedent for it in Semgrep, but there are a variety of
other partial programs that are allowed as patterns, so the
infrastructure for this exists.

However, we currently use the JS pfff parser to parse all patterns for
TS and JS, rather than this tree sitter parser which we use for parsing
targets (for JS, we still fall back on the pfff parser, but for TS we
use this exclusively).

Because we generally seem to be moving away from pfff parsers, I was
reluctant to add new features to the pfff parser. Instead, I would like
to begin to fall back on the tree sitter parser when parsing patterns,
just like we do when parsing targets.

This PR adds parsing support for the semgrep ellipsis (`...`) and top
level switch case parsing to the TypeScript and TSX parsers. The
ellipsis is a table stakes feature, so I didn't want to omit it even
initially. The switch case provides the motivation for adding the
fallback logic (in an upcoming PR to Semgrep).

Test plan: Automated tests plus a manual test with my upcoming changes
to Semgrep, where it can successfully match using a standalone switch
case pattern.