fix circleCI, do not use xlarge resource
This seems to not be part of our plan anymore.
We get
```
Resource class docker for xlarge is not available for your project, or
is not a valid resource class. This message will often appear if the
pricing plan for this project does not support docker use.
```
error in circleCI dashboard

test plan:
wait for circleCI to kick in and hope for green check