chore: Update tree-sitter-rust (3/n)
Among other things, this pulls in some significant changes how macros
are handled. There were two major changes, months apart, and I wanted to
pull them both in at once, so this is a fairly major jump.

I also ran in to a weird issue parsing the Semgrep ellipsis that was
sorted out in a later commit, so I pulled in some more commits after the
macro changes as well.

Test plan: Automated tests, plus pulled this in to Semgrep and verified
that tests pass there.