semgrep-kotlin: Add 'constructor' as a simple identifier
To allow classes in newlines, the grammar was modified by adding a
partial_class_declaration rule. But, this breaks parsing when
constructor is used as an identifier. This commit fixes that by adding
constructor as a simple identifier. The change is made in the
semgrep-kotlin grammar and not the upstream grammar since the breakage
is only seen with the partial_class_declaration rule, and not otherwise.